+++ fixtures/babel-component/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @compiled/babel-component-fixture
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [47050f4]
+  - @compiled/react@0.10.3
+
 ## 0.0.11
 
 ### Patch Changes