+++ packages/parcel-transformer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @compiled/parcel-transformer
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [307bb83]
+- Updated dependencies [18dcdf8]
+  - @compiled/babel-plugin@0.13.0
+
 ## 0.7.1
 
 ### Patch Changes