+++ packages/webpack-loader/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @compiled/webpack-loader
 
+## 0.8.4
+
+### Patch Changes
+
+- Updated dependencies [47050f4]
+  - @compiled/react@0.10.3
+
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [307bb83]
+- Updated dependencies [18dcdf8]
+  - @compiled/babel-plugin@0.13.0
+
 ## 0.8.2
 
 ### Patch Changes